+++ tests/test-text-replace.php
@@ -51,6 +51,7 @@ class Text_Replace_Test extends WP_UnitTestCase {
 		remove_filter( 'c2c_text_replace_case_sensitive', '__return_false' );
 		remove_filter( 'c2c_text_replace_comments',       '__return_true' );
 		remove_filter( 'c2c_text_replace_filters',        array( $this, 'add_custom_filter' ) );
+		remove_filter( 'c2c_text_replace_third_party_filters', array( $this, 'add_custom_filter' ) );
 	}
 
 
@@ -73,6 +74,17 @@ class Text_Replace_Test extends WP_UnitTestCase {
 		return array(
 			array( 'get_comment_text' ),
 			array( 'get_comment_excerpt' ),
+		);
+	}
+
+	public static function get_third_party_filters() {
+		return array(
+			array( 'acf/format_value/type=text' ),
+			array( 'acf/format_value/type=textarea' ),
+			array( 'acf/format_value/type=url' ),
+			array( 'acf_the_content' ),
+			array( 'elementor/frontend/the_content' ),
+			array( 'elementor/widget/render_content' ),
 		);
 	}
 
@@ -446,10 +458,46 @@ class Text_Replace_Test extends WP_UnitTestCase {
 		$this->assertGreaterThan( 0, strpos( apply_filters( $filter, 'a :coffee2code:' ), $expected ) );
 	}
 
+	/**
+	 * @dataProvider get_third_party_filters
+	 */
+	public function test_replace_applies_to_third_party_filters( $filter ) {
+		$expected = $this->expected_text( ':coffee2code:' );
+
+		$this->assertEquals( 2, has_filter( $filter, array( c2c_TextReplace::get_instance(), 'text_replace' ) ) );
+		$this->assertGreaterThan( 0, strpos( apply_filters( $filter, 'a :coffee2code:' ), $expected ) );
+	}
+
+	public function test_third_party_filters_are_part_of_c2c_text_replace_filters() {
+		$filters = array_map(
+			function ( $x ) { return reset( $x ); },
+			array_merge(
+				$this->get_third_party_filters(),
+				$this->get_default_filters()
+			)
+		 );
+
+		add_filter( 'c2c_text_replace_filters', array( $this, 'capture_filter_value' ) );
+
+		c2c_TextReplace::get_instance()->register_filters(); // Plugins would typically register their filter before this originally fires
+
+		$this->assertSame( $filters, $this->captured_filter_value[ 'c2c_text_replace_filters' ] );
+	}
+
 	public function test_replace_applies_to_custom_filter_via_filter() {
 		$this->assertEquals( ':coffee2code:', apply_filters( 'custom_filter', ':coffee2code:' ) );
 
 		add_filter( 'c2c_text_replace_filters', array( $this, 'add_custom_filter' ) );
+
+		c2c_TextReplace::get_instance()->register_filters(); // Plugins would typically register their filter before this originally fires
+
+		$this->assertEquals( $this->expected_text( ':coffee2code:' ), apply_filters( 'custom_filter', ':coffee2code:' ) );
+	}
+
+	public function test_hover_applies_to_custom_third_party_filter_via_filter() {
+		$this->assertEquals( ':coffee2code:', apply_filters( 'custom_filter', ':coffee2code:' ) );
+
+		add_filter( 'c2c_text_replace_third_party_filters', array( $this, 'add_custom_filter' ) );
 
 		c2c_TextReplace::get_instance()->register_filters(); // Plugins would typically register their filter before this originally fires